clustering/requirements.txt
requests==2.28.1
py7zr==0.20.0
click==8.1.3
pyyaml==6.0
pandas_read_xml==0.3.1
